Xcode 8, Swift 3.0 syntax compatible
Project file converted to Xcode 8 compatible
Swift code has been converted to v3.0 syntax
Date-format code signatures changed to EU
minimum supported iOS is now 8
Fixed a bug which was storing Searching event as Departed event in app
db

+++ Example/ExampleObjectiveC/PIONotificationViewController.m
@@ -2,9 +2,10 @@
 //  PIONotificationViewController.m
 //  ExampleObjectiveC
 //
-//  Created by ParkTAG on 8/26/16.
+//  Created by zee-pk on 26/08/2016.
 //  Copyright (c) 2016 predict.io by ParkTAG GmbH. All rights reserved.
 //
+
 #import <UIKit/UIKit.h>
 #import "PIONotificationViewController.h"
 #import "AppDelegate.h"
@@ -99,7 +100,7 @@
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    id <NSFetchedResultsSectionInfo> sectionInfo = self.fetchedResultsController.sections[section];
+    id <NSFetchedResultsSectionInfo> sectionInfo = self.fetchedResultsController.sections[(NSUInteger) section];
     return sectionInfo.numberOfObjects;
 }
 
@@ -130,7 +131,7 @@
 - (void)configureCell:(UITableViewCell *)cell withEvent:(EventViaNotification *)event {
     enum PredictIOEventType eventType = (enum PredictIOEventType) event.type.integerValue;
     if (eventType == TransportMode) {
-        cell.textLabel.text = self.transportationModeLabels[event.mode.integerValue];
+        cell.textLabel.text = self.transportationModeLabels[(NSUInteger) event.mode.integerValue];
     } else {
         cell.textLabel.text = self.labels[eventType];
     }
